// website/WebContent/src/app/generated/api/service-requester/model/interviewResponse.ts
/**
 * Service Requester REST Interface
 * Rest interface to communicate with the service requester.
 *
 * OpenAPI spec version: 0.0.2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface InterviewResponse { 
    dropdownList?: Array<string>;
    interviewState?: InterviewResponse.InterviewStateEnum;
    question?: string;
    questionID?: string;
    questionType?: InterviewResponse.QuestionTypeEnum;
}
export namespace InterviewResponse {
    export type InterviewStateEnum = 'MORE_INFORMATION_NEEDED' | 'INTERVIEW_DONE';
    export const InterviewStateEnum = {
        MOREINFORMATIONNEEDED: 'MORE_INFORMATION_NEEDED' as InterviewStateEnum,
        INTERVIEWDONE: 'INTERVIEW_DONE' as InterviewStateEnum
    };
    export type QuestionTypeEnum = 'FILE' | 'NO_QUESTION' | 'DROPDOWN' | 'NUMBER' | 'TEXT';
    export const QuestionTypeEnum = {
        FILE: 'FILE' as QuestionTypeEnum,
        NOQUESTION: 'NO_QUESTION' as QuestionTypeEnum,
        DROPDOWN: 'DROPDOWN' as QuestionTypeEnum,
        NUMBER: 'NUMBER' as QuestionTypeEnum,
        TEXT: 'TEXT' as QuestionTypeEnum
    };
}